backend/requirements.txt
asgiref==3.5.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
decorator==5.0.7
defusedxml==0.7.1
Django==3.2
django-admin-rangefilter==0.8.2
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
fonttools==4.33.3
idna==3.3
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.4.2
matplotlib==3.5.2
matplotlib-inline==0.1.2
numpy==1.22.4
oauthlib==3.1.1
packaging==21.3
parso==0.8.2
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.18
pycparser==2.21
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
traitlets==5.0.5
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
wordcloud==1.8.1
